+++ spec/spec_helper.rb
@@ -27,5 +27,5 @@ RSpec.configure do |config|
   config.order = :random
   Kernel.srand config.seed
 
-  config.before(:each) { Anyway.env.clear }
+  config.after(:each) { Anyway.env.clear }
 end