+++ spec/spec_norails_helper.rb
@@ -22,5 +22,5 @@ RSpec.configure do |config|
   config.order = :random
   Kernel.srand config.seed
 
-  config.before(:each) { Anyway.env.clear }
+  config.after(:each) { Anyway.env.clear }
 end